reorg(measurements) renamed measures to measurements
