+++ src/operations/measurements/measureArea.js
