+++ src/operations/measurements/measureArea.test.js
